moving TitleScreen to screens.py
